Moving to patched version of SDWebImage
The original library does not force the decoding of GIF images on a background thread, so it has been patched to do so. This may potentially be a 'bad thing' but it certainly improves the scroll performance of images.

Also making sure we only start an image load if we're actually mounted to a superview.